+++ app/about/head.tsx
@@ -1,11 +1,11 @@
-import { HeadMeta } from "../components/HeadMeta";
-import DefaultTags from "./DefaultTags";
+import { HeadMeta } from "../../components/HeadMeta";
+import DefaultTags from "../DefaultTags";
 
 const meta = {
-	title: "Milovan Gudelj - Web developer / UI designer",
+	title: "Milovan Gudelj - About me",
 	description:
 		"I design and develop engaging websites and delightful digital experiences.",
-	url: "https://milovangudelj.com/",
+	url: "https://milovangudelj.com/about",
 	image: "https://milovangudelj.com/images/og-image.png",
 };
