+++ components/TopArtists/TopArtists.tsx
@@ -1,10 +1,12 @@
+"use client";
+
 import { ComponentProps } from "react";
 import useSWRImmutable from "swr/immutable";
 import { twMerge } from "tailwind-merge";
 
 import fetcher from "../../lib/fetcher";
 import { TopArtists } from "../../lib/types";
-import { WrappedList } from "../";
+import WrappedList from "../WrappedList/WrappedList";
 
 export const Artists = ({ className }: ComponentProps<"div">) => {
 	const { data } = useSWRImmutable<TopArtists>("/api/top-artists", fetcher);
@@ -20,3 +22,5 @@ export const Artists = ({ className }: ComponentProps<"div">) => {
 		</div>
 	);
 };
+
+export default Artists;
